+++ vck/src/androidMain/kotlin/at/asitplus/wallet/lib/DefaultZlibService.android.kt
@@ -1,0 +1,3 @@
+package at.asitplus.wallet.lib
+
+actual class DefaultZlibService actual constructor() : DefaultAndroidJvmZlibService()